Small fix. Implemented screens.

+++ migrations/2023-08-07-202851_create_screens_table/up.sql
@@ -2,7 +2,7 @@
 CREATE TABLE screens (
     screen_id Uuid PRIMARY KEY NOT NULL,
     screen_name TEXT NOT NULL,
-    price_per_time NUMERIC(10, 2) NOT NULL,
+    price_per_time double precision NOT NULL,
     characteristics TEXT NOT NULL,
     traffic INTEGER NOT NULL,
     business_id Uuid NOT NULL,
