+++ migrations/2023-08-07-204522_create_payments_table/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 CREATE TABLE payments (
     payment_id UUID PRIMARY KEY NOT NULL,
-    price NUMERIC(10, 2) NOT NULL,
+    price double precision NOT NULL,
     user_id UUID NOT NULL,
     order_id UUID NOT NULL,
     FOREIGN KEY(user_id) REFERENCES users (user_id),
